+++ src/Attachment.php
@@ -25,6 +25,7 @@ final class Attachment
     private $display;
     private $description;
     private $fileUrl;
+    private $content;
 
     /**
      * @param IRI              $usageType   The type of usage of this attachment
@@ -34,9 +35,15 @@ final class Attachment
      * @param LanguageMap      $display     Localized display name (title)
      * @param LanguageMap|null $description Localized description
      * @param IRL|null         $fileUrl     An IRL at which the attachment data can be retrieved
+     * @param string|null      $content     The raw attachment content, please note that the content is not validated against
+     *                                      the given SHA-2 hash
      */
-    public function __construct(IRI $usageType, $contentType, $length, $sha2, LanguageMap $display, LanguageMap $description = null, IRL $fileUrl = null)
+    public function __construct(IRI $usageType, $contentType, $length, $sha2, LanguageMap $display, LanguageMap $description = null, IRL $fileUrl = null, $content = null)
     {
+        if (null === $fileUrl && null === $content) {
+            throw new \InvalidArgumentException('An attachment cannot be created without a file URL or raw content data.');
+        }
+
         $this->usageType = $usageType;
         $this->contentType = $contentType;
         $this->length = $length;
@@ -44,6 +51,7 @@ final class Attachment
         $this->display = $display;
         $this->description = $description;
         $this->fileUrl = $fileUrl;
+        $this->content = $content;
     }
 
     public function getUsageType()
@@ -79,6 +87,11 @@ final class Attachment
     public function getFileUrl()
     {
         return $this->fileUrl;
+    }
+
+    public function getContent()
+    {
+        return $this->content;
     }
 
     public function equals(Attachment $attachment)